login

+++ model/UserMapper.php
@@ -46,9 +46,16 @@ class UserMapper
     public function delete(int $id): void
     {
         try {
-            $stmt = $this->database->connect()->prepare('DELETE FROM users WHERE idUser = :id;');
-            $stmt->bindParam(':id', $id, PDO::PARAM_INT);
-            $stmt->execute();
+
+
+
+            {
+
+
+                $stmt = $this->database->connect()->prepare('DELETE FROM users WHERE idUser = :id and idPermission != 1;');
+                $stmt->bindParam(':id', $id, PDO::PARAM_INT);
+                $stmt->execute();
+            }
         }
         catch(PDOException $e) {
             die();
